Changed peninsula test case to A-grid
The conversion of the peninsula test case from A-grid to C-grid
significantly deteriorated the quality of the particle trajectories.
Since we are not testing A-to-C grid conversion, I have rewritten the
peninsula test case to be fully A-grid.

If anything, this shows the RK method is already pretty grid-agnostic,
as it is fine with receiving an A-grid