Peninsula: Tweaks to the peninsula test setup to ensure delta(P) < 1.e-5
This changes the grid coordinates to always be half a cell dimension away
from 0 to avoid numerical inconsistencies at the boundaries. It also moves
the initial y range of particles away from the boundary, in order to
ensure that the delta(P) between initial and final particle position is
below 1.e-5.