Generic netCDF Grid Class
In process of building test script with an altered version of the
NEMOgrid class to take arbitrary dimension and field names supplied
externally.

Appear to be loading data correctly, but issues in actually advecting
particles in resulting grid, suggesting that the new grid object is not
what particle.py or kernel.py expect.