Added uniform vertical velocity field to the peninsula test-case
Made the peninsula test-case 3D, by adding a uniform vertical velocity
to the field (and making horizontal velocities uniform in depth).

Note that this is a very simple test case, and not physically realistic
as there is no mass conservation. However, it should work to test
integration of vertical velocities in PARCELS