Sampling of user-defined fields
Added in the functionality for custom user particle subclasses of
Particle or JITParticle to define “user_vars” and sampling kernels
which will record an arbitrary field to sample from other than lat and
lon.
Not sure if I have done this best way, i’ve tried to edit the current
structure as little as possible. Something not working properly with
JITParticle subclasses- suspect it has something to do with me not
allocating memory for the new user_vars ordereddict in the Particle
Class