ParticleSet: Always perform while-loops for the inner time loop (leap)
This is quite a wholesale change that undoes the special-casing for
execute_adaptive and now always performs the "adaptive loop on the inner
time leaps. It also cleans up the the time stepping parameter derivation
and renames the `output_interval` parameter to `interval`. Most notably,
the outer time loop in execute now only cares about the number of leaps
performed and leaves the inner loop to `Kernel.execute()`.