masking update for AdvectionRK4 method
Before interpolating to find the velocity at a given (lat, lon), we
check to see if (lat, lon) are in the grid domain. If they are in the
domain , we proceed to the velocity interpolation step. If they are
outside the domain, the particle has exited and the coordinates get
masked to np.nan. Note that np.nan masking is compatible with the show
movie feature of the execute method and the plotParticles 2d mode
script.