Write flag added to particle variables
Added a simple `to_write` flag to the custom variable definition
(default is true). `ParticleFile` objects now only write particle
variables to netcdf if this flag is on.

This saves on file size for internal particle variables that are used
at runtime, but have no need to be written to results.