Diffusion kernel and test
This commit adds a 2d diffusion kernel for particles in spatial varying
diffusivity fields.

A test file is included demonstrating how brownian motion results in
erroneous concentrations of particles in areas of low diffusivity.

Not sure if we wish to add a new kernel file such as `diffusion.py`,
but please let me know your comments.