Moving initial writing of particle file to .execute loop
Removing the initial_dump keyword from ParticleFile __init__, as
adding/deleting of particles can interact adversely with that

Now, particleFile is updated before kernel execution. Advantages are
- ‘Time’ of particles is knows, so doesn’t have to be guessed as 0.
- Particles can be added/deleted in between .execute calls
- Start locations of particles are always accurately written

Disadvantage of this implementation is that the last particle position
will not be written. Since we never know when will be the ‘last’ one
(as there can always be another call to .execute, I think the only
solution is that users who want to also write the last particle
positions do a final call to particle_file.write() at the end of their
script