Adding add_periodic_halo method to grid and Field
Implementation for adding a ‘periodic halo’ to each Field in a grid.
The halo is made through extending the Field (and lon/lat) by copying a
small portion of the field on one side of the domain to the other.
Halo can be added in zonal, meridional, or both directions
Default halo size is 3% of Field dimension