Removing grid.time and grid.depth
Since `grids` are simply a collection of `fields`, it doesn’t make
sense to have both a `grid.time` and a `grid.`(field)`.time`. Same for
`grid.depth` and `grid.`(field)`.depth`. We already store lon and lat
on the fields only, the same should be done with depth and time.

This also allows for easier mixing of fields of different
shapes/dimensions (e.g. #41)