Adding boolean allow_time_extrapolation to Field
This will allow more control over whether to throw an error when
evaluating fields outside of the time range of the Field.

Set default to True for Grid.from_data. And False for Grid.from_netcdf.
Also default True if there is no time dimension on grid.

Currently only implemented in scipy