Renaming `trajectory` to `id` in Particle Variables
As suggested by @mlange05

Note that, due to a cptr bug in particle.py, calling the Variable `xid`
for now.

It seems that the variable name needs to be alphabetically after
`time`, otherwise JIT mode goes wonky. So `trajectory` and `xid` work,
but `id` doesn’t.
Likely a padding/misalignment issue with JIT; to be fixed…