Fixed a bug in JIT ParticleSet.remove
Seems that particles need to be deleted before updating _cptr in JIT
mode. Otherwise, ordering goes wrong. This wasn’t spotted before since
we only deleted the last particle

Now added a unit test to test_particle_sets where a particle somewhere
in the middle of the ParticleFile is removed. Also use this test to
check array-style writing of ParticleFile