Allowing for advance time to update only a subset of the fields
There are use cases when not all fields in a FieldSet need to be
updated by advance time (e.g. if some are time-independent). Then, it
makes more sense to update only those fields that are in the new
FieldSet, rather than all fields that are in the old FieldSet