Install newer Anaconda (AD5) compiler
This drops the dependency on `gcc` in favor of the new compiler tools
(<https://conda.io/docs/user-guide/tasks/build-packages/compiler-tools.html>).
As these new packages differ for Linux and OSX, we'll also have to give
up a common environment file for Linux and OSX.  This is already taken
into account in the documentation.