Set python version when copying example notebooks
The notebooks included in the examples usually contain metadata specifying the
kernel to use.  For notebooks saved after using a Python 2 kernel, this will
ensure that Jupyter tries to pick the correct kernel when the notebook is
opened again.  For testing with `pytest --nbval*`, however, this means that
nbval will fail if a notebook specifying Python 2 is evaluated from a Python 3
environment.

This commit adds a function to `parcels_get_examples` which reads the current
Python version from `sys` and then writes a corresponding `kernelspec` to the
notebook metadata.