Allowing for plotting 'land=True' without field='vector'
This addresses #330. Note however, that `land=True` now only works if
either now field is given, or if `field=‘vector’`. In all other cases,
the `field` is simply not shown

Too fix that follow-up bug is not trivial at all; and would probably
require moving away from Basemap and use Cartopy instead, as was the
plan anyways (see #221)