Calculating distance updated
In the Calculating distance travelled part of the parcels_tutorial, I changed the example from one with particles in the MovingEddies fieldset to one with particles in a GlobCurrent fieldset, which actually measures the latitude and longitude in degrees North/East. I also shortly mentioned that some datasets measure them in (k)m, in which case the function TotalDistance must be different.