Use weakref.finalize in Kernel
To avoid accidentally keeping the Kernel alive and preventing garbage
collection, we can't own any references to the Kernel in the
finalizer. The cleanup routines need to live outside the object, and
we register the required arguments on compile and library load for the
file cleanup and library unload finalizers, respectively.