Changing Error to 0
Instead of throwing errors when the point to evaluate is on land, I suggest to return 0. This for the simple reason that user might want to evaluate points that aren't at the particle location (for gradients for example). Also, the tracer land mask might be different that the (U,V) land mask, which implies that a particle can be in the ocean for (U,V) but not for a tracer, and execution shouldn't stop for this.