Use .data in Field from xarray
Using .values to extract the data from an xarray dataset when creating
a field forcing loading of all the data. This doesn't take advantage
of xarray datasets with deferred data. Indeed, the rest of Field is
aware of dask arrays, so we can load them in directly by using .data
instead.