Support chunked loading from xarray datasets
This allows the use of the deferred/chunked loading mechanism on
xarray datasets. In particular, this may be useful for dask-backed
datasets which are not loaded directly from files (e.g. some
transformation/chunking/etc. is performed on the data).

In its current form, this will break the usual deferred arrays, which
assume field.data only contains 3 elements along the first axis --
using xarray datasets, this is now the full time dimension and must be
indexed accordingly.