Support sampling fields without applying conversion
Velocity fields on a spherical grid will automatically be converted to
degrees when sampled within a kernel. This makes sense for advection,
but we may want an interpolated sampling, without unit conversion. By
calling field.eval(..., applyConversion=False) directly, we can get
the desired behaviour. This works in both scipy and JIT modes.