Added unit test for uneven c-grids, where UVEL[nx+1, ny] and VVEL[nx, ny+1].
Standard netCDF output generated through MITgcm's `mnc` module uses this type of grid,
whereas NEMO has velocity output in the form U[nx+1, ny+1] and V[nx+1, ny+1]. The
first row in U and the first column in V is zero in these cases. When MITgcm data is subsetted,
it often also has this square shape, with the last row and column respectively not used.