Refactor C code for structured grid temporal interpolation
There's a lot of code duplication here which hides the actual
intent.

This was also hiding a behavioural difference where
interpolation on a 2D field with CGRID_VELOCITY would cause an error
only if the interpolation time exactly coincides with the time
grid. The new behaviour is to not raise the error, regardless of the
interpolation time.