Fixing a bug when Kernels with multiple docstrings could not be generated
As reported in #910, the ast fix for python3.8 in #753 only appeared to work when the docstring was on the top of the concatinated kernel.

Now, we assume that all strings are docstrings or comments by default and can be removed. The only exception is the string 'parcels_customed_Cfunc_pointer_args', which is needed when c kernels are included (see tests/test_kernel_language.py -> test_c_kernel)

This fixes #910