Remove explicit loops over particle state
The loops to test and set particle state were implemented using Python
primitives, which are extremely slow for large particle counts. We can
easily modify remove_deleted_soa, because we know the underlying
collection supports array indexing. We should also be able to rely on
Collection.__getitem__ returning something with an array-like
interface that we can use to set all particle states simultaneously:
for SoA the array assignment is significantly faster, and it would be
no slower for other collections.