Update tutorial_particle_field_interaction.ipynb
While doing further work on field-particle interaction with @erikvansebille and @dlobelle, I noticed that accessing the field value through fieldset.C[particle] and fieldset.C[particle.ti, particle.yi, particle.xi] returns different results. This is because the former (field evaluation) interpolates the field at particle location. While this seems intuitively convenient, this doesn't allow to close the mass balance between the particle and the cell that contains it, that has to be updated through indexing instead.
This mismatch and potential consequent issues have been  briefly explained in this updated version of the Particle-Field Interaction Notebook.