Rewriting read_from_npy in particlefilesoa so that it creates smaller files
Previously, an array was created with the total number of unique timesteps, which could become very large when lots of particles were deleted at different times. Access rows and columns were later deleted. This has now been changed so that the initial data array is much smaller and less likely to get out of Memory errors