Raising NotTestedError for some of the methods in collectionsoa and collectionaos that are not tested in unit-tests
And commenting out the codes for these. Will be available when we need these codes for e.g. dynamic load-balancing in MPI. For now, however, they do not seem to be used or tested and hence we can't guarantee their integrity. Creating test-cases for them would take to much effort now