Update particleset.py
When you want to populate a particle set prior to a parcels run, this can be done with the help of a KDTree search. However, when the input fields have missing values (i.e.: nan's) in their grid, the algorithm does provide an answer, but these are not correct. Hence, the following adjustments are made, to remove nan entries from the creation of a KDTree.