[#1874] Initial draft of reduction to single index
Temporal indexing is left separate from spatial indexing due to the
fact that ti was initialized differently from spatial indices on
L131 of particleset.py (ti = -1, xi,yi,zi=0). To stick with this
convention and to avoid an unforeseen consquences of merging spatial and
temporal indexing, I've opted to kep them separate with ei=0 initially
and ti=-1 initially.

ravel_index/unravel_index methods have been added to the Field class.

In this commit, some tests aren't passing as it appears some tests
attempt to check the values of the old xi,yi,zi .