Add dimrange function and gridset_size property to fieldset
The fieldset.dimrange function is meant to be a replacement for
fieldset.grid.dimrange. Functionally, it operates in the same way but
relies on the fields underneath the (u)xarray.(Ux)dataset. Currently,
this is incompatible with a fieldset constructed using the
`fieldset.add_field` function, since add_field does not update the `ds`.
Need to think on how we want to track this.

The `fieldset.gridset_size` property is set to be equivalent to the
number of fields (dataarrays) stored under the fieldset. This property
is meant to replace the `fieldset.gridset.size` property.