Updating Parcels pitch in readme
As we move towards v4, I feel it's time to say goodbye to the funny bacronym for Parcels. We don't use it anymore on the revamped oceanparcels.org website, and when we remove it here in the readme it will be gone from the codebase too.

I also aligned the pitch with that on the oceanparcels website