Clean up init for list of fields and fix v4 test errors
This commit also brings in changes to add_field and add_vector_field to
avoid using the setattr method. Instead, we opt to extend the dictionary
for the fields (which map {field.name : field}). The add_vector_field
method optionally adds the VectorField object (if it's not already
present) and each of its components by calling add_field