Remove allow_time_extrapolation attribute on Field
When a xarray or uxarray dataarray does not have a time dimension, one
is added to the Field.data dataarray. When a field.data datarray has a
time dimension that is length 1, we assume that extrapolation is
allowed meaning that we just keep using the same single time-slice for
that field.