Create xgcm submodule
Since we're vendoring code from XGCM here, it's nice to have a clear separation of what is parcels code and what is XGCM code in case in future we are able to directly depend on XGCM itself (removing the vendored code). That is if the development goals of XGCM are aligned with Parcels (i.e., they are willing to expose grid geometry and axis information in their public API). This separation is also helpful because it lines up well with the unstructured case, where we depend on the UXarray grid and extend its functionality as well via composition.