Add test for constant in time/no time interval field
This test verifies that a field that has no time interval (is constant
in time) evaluates to the same value. Additionally, this test is a
minimal reproducer for the case that the patch in the
search_time_indices function is meant to resolve.

The search_time_indices function is called in the `Field.eval` method, which
always expects a tuple output corresponding to the time index and the
time interpolation weight. Previously, when there was no time interval,
the search_time_indices function only returned a single value.