Add grid pointer in vectorfield, clean up vectorfield.eval logic
For consistency in the API between a field and vectorfield, the
VectorField.grid points to VectorField.U.grid . This is needed since
both Field and VectorField objects can be used in a FieldSet and it
would be cumbersome to have logic switch how a `grid` is obtained based
on whether a fieldset attribute is a Field or VectorField.

VectorField.eval logic now follows a similar execution pattern to the
Field.eval logic, except when the VectorField._interp_method is not
provided. When VectorField._interp_method is not provided, interpolation
falls back to component-wise interpolation using the
VectorField.[U,V,W]._interp_method's.