Remove generating and storing kernel code in pset.execute
Looks like the implementation doesn't work (old code from v3) and also isn't very maintainable.

This was to support seeing if kernels had changed, and then recompiling them so that pset.execute wasn't slow.

This can be implemented in v4 in a maintainable way - but this isn't a priority for now.