Update search to handle array of lat,lon
To handle array of lat,lon the x and y values need to be column stacked.
After making this change, the internal API funciton in uxgrid
(grid.neighbors._triangle_area) fails due to an assumed shape for the
coordinate (being a scalar and not an array). The quick workaround here
is to bring the `_barycentric_coordinates` method into parcels.uxgrid
and leverage the numpy-ified parcels.uxgrid._triangle_area method we
already have defined.

In making this change, since order of operations is different for the
barycentric coordinate calculation, this required updating the rounded
and hashed values in the stommel gyre test. The actual values for the
lat/lon still appear to be ok.